Deploy to Production Fixes Various fixes are needed to deploy to production. Update requirements, drop database name, etc.
Remove Database Name from Queries
The database name doesn't need to be in the queries.

One More Database Name

Add uWSGI to Requirements
uWSGI is used on the production server.

Add Startup Script for uWSGI
A basic script to run uWSGI in the background and be able to kill it.

Add Usage Information for Index GET
Usage information and other information about everest are returned
on a GET to /

Add NetworkX to Requirements
Forgot to update requirements.txt last time around.

Remove Database Name from Queries
The database name doesn't need to be in the queries.

One More Database Name

Add uWSGI to Requirements
uWSGI is used on the production server.

Add Startup Script for uWSGI
A basic script to run uWSGI in the background and be able to kill it.