Revamp Categories This change completely revamps the way categories are handles. Two main differences: multiple categories can be selected via POST and the category list is passed around as a list instead of an int.
/industry/ accepts POST now. It takes a list under "categories",
bool for "rigs", and bool for "names". The GET endpoint is also
changed a bit. /industry/all/ and /industry/norigs/ have been removed.
Everything should use either the POST or /industry/<cat id>/ with
optional /norigs/ and /names/ at the end (in any order). Cat ID can
be 0 for all.

Internally, to support multiple categories instead of just all
categories, an array of categories is passed around. This is joined
together and used in "categoryID in (val)".